Move all editorial pages under the reserved /-/ namespace
"-" can never be a GitHub login (logins can't start or end with a hyphen), so
content under /-/ can never shadow the single-segment $user route — no more
per-page "is this word a safe username?" policy debates. The metrics hub may
now own its bare prefix, so /metrics/explained becomes /-/metrics.

Old URLs (/sponsoring, /metrics/explained, /metrics/<slug>,
/organizations/<slug>) 301 to their new homes via thin redirect routes;
/company/<slug> now points straight at /-/organizations/<slug> to avoid a
double hop. Sitemap, canonicals, OG urls, JSON-LD breadcrumbs, and all
internal links (routes, components, MDX articles) updated. OG image assets
keep their historical og/metrics/ paths.
