+++ src/content/metrics/commits.mdx
@@ -25,7 +25,7 @@ only when all of these hold:
   count after they're merged upstream.
 
 The count we show is the public one; work in private repositories shows up
-separately as [private contributions](/metrics/private-contributions), if the
+separately as [private contributions](/-/metrics/private-contributions), if the
 user has enabled that.
 
 ## Why workflow dominates the number
@@ -42,8 +42,8 @@ How a team merges is worth more commits than how much it codes:
 
 So a huge commit count can mean a prolific engineer — or a granular committer,
 a monorepo bot operator, or someone whose repo gets an automated commit every
-night. Cross-check it against [pull requests](/metrics/pull-requests) and
-[reviews](/metrics/reviews): a human pattern usually shows activity across
+night. Cross-check it against [pull requests](/-/metrics/pull-requests) and
+[reviews](/-/metrics/reviews): a human pattern usually shows activity across
 several types.
 
 ## What it doesn't measure
