+++ src/content/metrics/followers.mdx
@@ -15,7 +15,7 @@ worth having next to the others.
 
 The current follower count on a user's GitHub profile — a snapshot, not a
 history. GitHub doesn't expose when each follow happened, so unlike
-[commits](/metrics/commits) there's no lifetime curve to draw; followers
+[commits](/-/metrics/commits) there's no lifetime curve to draw; followers
 appear on the [leaderboard](/?metric=followers) but have no chart on profile
 pages. We refresh the count when a profile's data is refreshed.
 
@@ -32,7 +32,7 @@ Follower counts obey attention dynamics, not effort dynamics:
 
 The result is that followers and contribution counts are **nearly independent
 axes**. There are engineers with tens of thousands of
-[total contributions](/metrics/total-contributions) and three hundred
+[total contributions](/-/metrics/total-contributions) and three hundred
 followers, and developer-educators with the inverse. Neither number debunks
 the other; together they locate someone on a builder ↔ broadcaster map.
 
@@ -47,7 +47,7 @@ counts as differences in audience, not ability.
 ## Why we include it anyway
 
 Because the contrast is informative. Sorting the same population of
-developers by followers and then by [reviews](/metrics/reviews) produces
+developers by followers and then by [reviews](/-/metrics/reviews) produces
 almost comically different leaderboards — one is who you've *heard of*, the
 other is who keeps the code flowing. Having both on one site is the point.
