+++ src/content/metrics/issues.mdx
@@ -16,7 +16,7 @@ your issue fixed.
 - **Opening an issue** in a public repository: +1.
 - Commenting on issues, closing them, or being assigned: not counted.
 - Issues in private repositories: absorbed into
-  [private contributions](/metrics/private-contributions).
+  [private contributions](/-/metrics/private-contributions).
 
 That narrowness makes the metric easy to read literally: it measures how often
 someone *starts* a written conversation about a problem.
@@ -35,7 +35,7 @@ High lifetime issue counts cluster around a few recognisable profiles:
   issues rather than commits.
 
 That's the key insight: **a high issue count with a modest
-[commit count](/metrics/commits) is not a red flag.** It often marks someone
+[commit count](/-/metrics/commits) is not a red flag.** It often marks someone
 whose contribution is finding and describing problems precisely — which is
 frequently harder than fixing them.
 
@@ -45,7 +45,7 @@ Quality of the reports, whether they were valid, or whether they were ever
 resolved. A hundred careful reproductions and a hundred "doesn't work, plz
 fix" weigh the same. It also misses all the issue *work* that isn't opening:
 triaging, deduplicating, and answering — labor that currently has no number
-anywhere on GitHub except partially in [reviews](/metrics/reviews).
+anywhere on GitHub except partially in [reviews](/-/metrics/reviews).
 
 ## How commit-history.com tracks it
 
@@ -57,5 +57,5 @@ someone becomes a maintainer and the issue tracker becomes their desk.
 ## See it in action
 
 [Look up a maintainer you know](/) and switch the chart to **Issues** — then
-check their [Total](/metrics/total-contributions) to see how the types stack
+check their [Total](/-/metrics/total-contributions) to see how the types stack
 together.
