+++ src/content/metrics/pull-requests.mdx
@@ -15,12 +15,12 @@ the day it was opened.
 
 - **Opening a pull request** in a public repository: +1, immediately.
 - Getting it merged: no extra credit — the merge shows up in the *author's*
-  [commit count](/metrics/commits) instead (if it lands on the default
+  [commit count](/-/metrics/commits) instead (if it lands on the default
   branch).
 - Reviewing someone else's PR: counted separately, under
-  [reviews](/metrics/reviews).
+  [reviews](/-/metrics/reviews).
 - PRs in private repositories: folded into the opaque
-  [private contributions](/metrics/private-contributions) count.
+  [private contributions](/-/metrics/private-contributions) count.
 
 ## Reading the number
