+++ src/content/metrics/repositories.mdx
@@ -18,7 +18,7 @@ match what their profile page shows.
   forkers can show a much lower number here than their profile suggests.
 - Repositories created inside private accounts or orgs you can't see:
   invisible here (creation events in private contexts are part of
-  [private contributions](/metrics/private-contributions)).
+  [private contributions](/-/metrics/private-contributions)).
 
 This is also why our number can differ from the repository count on a GitHub
 profile: the profile's "Repositories" tab shows what a user currently *owns*
@@ -40,15 +40,15 @@ one. It tends to mark:
 The inverse profile is just as meaningful: some of the most prolific
 engineers on GitHub have a *tiny* repo count, because a decade of their work
 lives inside two or three big projects. Check the
-[commit count](/metrics/commits) next to it — thousands of commits across few
+[commit count](/-/metrics/commits) next to it — thousands of commits across few
 repos reads very differently from few commits across a thousand repos.
 
 ## What it doesn't measure
 
 Whether anything happened after `git init`. Creating a repository is the
 cheapest contribution GitHub counts — it weighs the same as a commit or a
-[review](/metrics/reviews) in the
-[Total](/metrics/total-contributions) sum, while representing the least
+[review](/-/metrics/reviews) in the
+[Total](/-/metrics/total-contributions) sum, while representing the least
 guaranteed work.
 
 ## How commit-history.com tracks it
