+++ src/content/metrics/reviews.mdx
@@ -18,7 +18,7 @@ changes, or submits a review comment on someone else's PR.
 - Plain inline comments outside a review, issue discussions, or mentoring in
   chat: not counted.
 - Reviews in private repositories: folded into
-  [private contributions](/metrics/private-contributions).
+  [private contributions](/-/metrics/private-contributions).
 
 ## Why this number matters more than it looks
 
@@ -26,8 +26,8 @@ Reviewing is the classic case of engineering work that's essential and
 invisible. A thorough review of a complex change can take longer than writing
 it — reading the diff, checking the edge cases, testing locally, writing
 feedback that lands well. None of that produces a commit under the reviewer's
-name. The author gets the [commit](/metrics/commits) and the
-[PR](/metrics/pull-requests); the reviewer gets… this number.
+name. The author gets the [commit](/-/metrics/commits) and the
+[PR](/-/metrics/pull-requests); the reviewer gets… this number.
 
 That's why review counts skew toward a specific profile: **senior engineers
 and maintainers**. As people grow in a codebase, their output shifts from
@@ -56,5 +56,5 @@ being mostly-author and became mostly-editor.
 
 [Look up a profile](/) and switch to Reviews, or
 [compare an author-heavy and a reviewer-heavy developer](/torvalds,gaearon) —
-then check [Total](/metrics/total-contributions), which is where
+then check [Total](/-/metrics/total-contributions), which is where
 review-heavy engineers finally get full credit.
