+++ src/content/organizations/members.mdx
@@ -22,7 +22,7 @@ see that you're a member at all.
 GitHub only lets *you* do this for yourself — an org admin can't publish your
 membership for you. Once it's public, you'll appear the next time the
 organization's data is refreshed, and your contributions start counting toward
-the [organization totals](/organizations/stats) too.
+the [organization totals](/-/organizations/stats) too.
 
 ## Other reasons someone is missing
 
@@ -39,6 +39,6 @@ the [organization totals](/organizations/stats) too.
 
 Being on the list with a **0** is normal and not a bug. The numbers are
 contributions **to this organization's public repositories** only — see
-[how organization stats are calculated](/organizations/stats). A member shows
+[how organization stats are calculated](/-/organizations/stats). A member shows
 zero when their work in the org happens in private repositories, under an email
 not linked to their GitHub account, or simply in other places entirely.
