+++ src/content/organizations/stats.mdx
@@ -36,7 +36,7 @@ things are excluded, and they can make the totals smaller than you'd expect:
 - **Private members.** GitHub keeps organization membership private by
   default, and we can only see members who made theirs public. An active
   organization whose people all keep membership private shows a total of zero —
-  see [why am I not on this member list?](/organizations/members).
+  see [why am I not on this member list?](/-/organizations/members).
 - **Outside contributors.** Someone who sends pull requests to the org's
   repos without being a member isn't counted. The measure is about the
   organization's people, not its repos' traffic.
